Add missing push opcodes
Added a new fixure, this one the json result from the remix compiler chewing on counter.sol.

Added all missing push opcodes.

This still crashes with "bad opcode" because there appears to be a section at the end of the opcode stream which is not valid opcodes.  Not clear why yet.

Replaced ArgValue with Option<UInt256> for less typing.  I could still remove the giant take_arg match if I could just carry the number of bytes in the ArgType somehow.