Restructure execute in prep for wider usage
Also added support for all Push instructions using some fancy rust matching. :heart:

This (probably needlessly) now copies the instructions out of a stream into a vector for execution?  I probably can remove that.  Unclear.

We probably want to add some sort of VM type which holds the stack and the InputManager and soon the Memory.