Implement enough to hit a different error-return
Now we don't end with a REVERT but rather with a RETURN which contains all of the bytecode.   Presumably we're hitting another error condition (like lack of message name perhaps?)

Had to add JUMPDEST, POP, RETURN and CODECOPY to make it run this far.

I worry I'm violating some rust-isms with things like to_usize_range.

Also now pass the message into Task::new so that value (the 'wei' ammount?) is non-zero and thus we can get to this error condition. :)