Execute the returned contract from the wrapper
This restructures execution, including using @abarth's suggestion of an InstructionResult and TaskResult to contain the return data rather than holding it on the Task object.

This is all in service of making it possible to return the contract from running the wrapper the first time and then be able to start executing that returned contract.

I'm not really sure I like how the execution flow is structured, but I suspect I don't have the right rust-smell for this yet.

Message is also borrowed with a lifetime.  I don't really understand everything I was typing, so it's possible I did it wrong.  Eventually message probably needs to be borrowed mutably and then its gas edited?  Not sure?