Add CALLDATASIZE and LT
Added CALLDATASIZE and LT and the necessary conversion functions to UInt256 as well as PartialOrd trait support as well as unit tests.  (If the members of UInt256 got reordered PartialOrd support would break.)

Change error type of send_message_to_contract to a new ContractError which includes Revert as a value and treat Reverts as errors (for now at least).