Add CALLDATALOAD and add a fake message
Implemented CALLDATALOAD, which then hit an out-of-bounds read error as it was trying to read past the end of the passed in message data.  So for now I've filled the message data with 0s, and some attempt at the first four bytes as an encoding (but it's not clear what I've done is correct).